requirements.txt
streamlit
requests
pandas
numpy
PyPDF2
python-docx
rich
groq
python-dotenv
youtube-transcript-api
Pillow